Move flat files and dockerCompose to src
