+++ commands/showlogs-container.ts
@@ -6,7 +6,7 @@
 import { IActionContext } from 'vscode-azureextensionui';
 import { ContainerNode } from '../explorer/models/containerNode';
 import { RootNode } from '../explorer/models/rootNode';
-import { ext } from '../extensionVariables';
+import { ext } from '../src/extensionVariables';
 import { AllStatusFilter, ListContainerDescOptions } from './utils/docker-endpoint';
 import { quickPickContainer } from './utils/quick-pick-container';
