+++ commands/utils/quick-pick-azure.ts
@@ -8,10 +8,10 @@ import { ResourceGroup } from 'azure-arm-resource/lib/resource/models';
 import { Location, Subscription } from 'azure-arm-resource/lib/subscription/models';
 import * as vscode from "vscode";
 import { IAzureQuickPickItem, UserCancelledError } from 'vscode-azureextensionui';
-import { skus } from '../../constants';
 import { openExternal } from '../../explorer/utils/openExternal';
-import { ext } from '../../extensionVariables';
 import { ResourceManagementClient } from '../../node_modules/azure-arm-resource';
+import { skus } from '../../src/constants';
+import { ext } from '../../src/extensionVariables';
 import * as acrTools from '../../src/utils/Azure/acrTools';
 import { isValidAzureName } from '../../src/utils/Azure/common';
 import { AzureImage } from "../../src/utils/Azure/models/image";
