+++ configureWorkspace/config-utils.ts
@@ -5,7 +5,7 @@
 
 import vscode = require('vscode');
 import { IAzureQuickPickItem } from 'vscode-azureextensionui';
-import { ext } from "../extensionVariables";
+import { ext } from "../src/extensionVariables";
 import { Platform, PlatformOS } from '../src/utils/platform';
 
 /**
