+++ explorer/models/commonRegistryUtils.ts
@@ -6,8 +6,8 @@
 import * as moment from 'moment';
 import * as vscode from 'vscode';
 import { parseError } from 'vscode-azureextensionui';
-import { MAX_CONCURRENT_REQUESTS, PAGE_SIZE } from '../../constants'
-import { ext } from '../../extensionVariables';
+import { MAX_CONCURRENT_REQUESTS, PAGE_SIZE } from '../../src/constants'
+import { ext } from '../../src/extensionVariables';
 import { AsyncPool } from '../../src/utils/asyncpool';
 import { Manifest, ManifestHistoryV1Compatibility } from '../utils/dockerHubUtils';
