+++ explorer/models/dockerHubNodes.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext } from 'vscode-azureextensionui';
-import { MAX_CONCURRENT_REQUESTS } from '../../constants';
+import { MAX_CONCURRENT_REQUESTS } from '../../src/constants';
 import { AsyncPool } from '../../src/utils/asyncpool';
 import { treeUtils } from '../../src/utils/treeUtils';
 import * as dockerHub from '../utils/dockerHubUtils';
