+++ explorer/models/registryRootNode.ts
@@ -8,8 +8,8 @@ import * as ContainerModels from 'azure-arm-containerregistry/lib/models';
 import { SubscriptionModels } from 'azure-arm-resource';
 import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext, parseError } from 'vscode-azureextensionui';
-import { keytarConstants, MAX_CONCURRENT_REQUESTS, MAX_CONCURRENT_SUBSCRIPTON_REQUESTS } from '../../constants';
-import { ext } from '../../extensionVariables';
+import { keytarConstants, MAX_CONCURRENT_REQUESTS, MAX_CONCURRENT_SUBSCRIPTON_REQUESTS } from '../../src/constants';
+import { ext } from '../../src/extensionVariables';
 import { AsyncPool } from '../../src/utils/asyncpool';
 import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
 import { getLoginServer } from '../../src/utils/nonNull';
