+++ main.js
@@ -17,7 +17,7 @@ let perfStats = {
 Object.defineProperty(exports, "__esModule", { value: true });
 
 const ignoreBundle = !/^(false|0)?$/i.test(process.env.AZCODE_DOCKER_IGNORE_BUNDLE || '');
-const extensionPath = ignoreBundle ? "./out/extension" : "./dist/extension.bundle";
+const extensionPath = ignoreBundle ? "./out/src/extension" : "./dist/extension.bundle";
 const extension = require(extensionPath);
 
 async function activate(ctx) {
