+++ src/dockerHubSearch.ts
@@ -6,7 +6,7 @@
 'use strict';
 
 import https = require('https');
-import { httpsRequest } from './src/utils/httpRequest';
+import { httpsRequest } from './utils/httpRequest';
 
 export function tagsForImage(image: IHubSearchResponseResult): string {
     let tags: string[] = [];
