+++ src/dockerInspect.ts
