+++ src/extension.ts
@@ -12,58 +12,58 @@ import { RequestPromise } from 'request-promise-native';
 import * as vscode from 'vscode';
 import { AzureUserInput, callWithTelemetryAndErrorHandling, createTelemetryReporter, IActionContext, registerCommand as uiRegisterCommand, registerUIExtensionVariables, TelemetryProperties, UserCancelledError } from 'vscode-azureextensionui';
 import { ConfigurationParams, DidChangeConfigurationNotification, DocumentSelector, LanguageClient, LanguageClientOptions, Middleware, ServerOptions, TransportKind } from 'vscode-languageclient/lib/main';
-import { viewACRLogs } from "./commands/azureCommands/acr-logs";
-import { LogContentProvider } from "./commands/azureCommands/acr-logs-utils/logFileManager";
-import { createRegistry } from './commands/azureCommands/create-registry';
-import { deleteAzureImage, untagAzureImage } from './commands/azureCommands/delete-image';
-import { deleteAzureRegistry } from './commands/azureCommands/delete-registry';
-import { deleteRepository } from './commands/azureCommands/delete-repository';
-import { pullImageFromAzure, pullRepoFromAzure } from './commands/azureCommands/pull-from-azure';
-import { quickBuild } from "./commands/azureCommands/quick-build";
-import { runTask, runTaskFile } from "./commands/azureCommands/run-task";
-import { showTaskProperties } from "./commands/azureCommands/show-task";
-import { TaskContentProvider } from "./commands/azureCommands/task-utils/showTaskManager";
-import { buildImage } from './commands/build-image';
-import { composeDown, composeRestart, composeUp } from './commands/docker-compose';
-import inspectImage from './commands/inspect-image';
-import { openShellContainer } from './commands/open-shell-container';
-import { pushImage } from './commands/push-image';
-import { consolidateDefaultRegistrySettings, setRegistryAsDefault } from './commands/registrySettings';
-import { removeContainer } from './commands/remove-container';
-import { removeImage } from './commands/remove-image';
-import { restartContainer } from './commands/restart-container';
-import { showLogsContainer } from './commands/showlogs-container';
-import { startAzureCLI, startContainer, startContainerInteractive } from './commands/start-container';
-import { stopContainer } from './commands/stop-container';
-import { systemPrune } from './commands/system-prune';
-import { tagImage } from './commands/tag-image';
-import { docker } from './commands/utils/docker-endpoint';
-import { DefaultTerminalProvider } from './commands/utils/TerminalProvider';
-import { DockerDebugConfigProvider } from './configureWorkspace/configDebugProvider';
-import { configure, configureApi } from './configureWorkspace/configure';
+import { viewACRLogs } from "../commands/azureCommands/acr-logs";
+import { LogContentProvider } from "../commands/azureCommands/acr-logs-utils/logFileManager";
+import { createRegistry } from '../commands/azureCommands/create-registry';
+import { deleteAzureImage, untagAzureImage } from '../commands/azureCommands/delete-image';
+import { deleteAzureRegistry } from '../commands/azureCommands/delete-registry';
+import { deleteRepository } from '../commands/azureCommands/delete-repository';
+import { pullImageFromAzure, pullRepoFromAzure } from '../commands/azureCommands/pull-from-azure';
+import { quickBuild } from "../commands/azureCommands/quick-build";
+import { runTask, runTaskFile } from "../commands/azureCommands/run-task";
+import { showTaskProperties } from "../commands/azureCommands/show-task";
+import { TaskContentProvider } from "../commands/azureCommands/task-utils/showTaskManager";
+import { buildImage } from '../commands/build-image';
+import { composeDown, composeRestart, composeUp } from '../commands/docker-compose';
+import inspectImage from '../commands/inspect-image';
+import { openShellContainer } from '../commands/open-shell-container';
+import { pushImage } from '../commands/push-image';
+import { consolidateDefaultRegistrySettings, setRegistryAsDefault } from '../commands/registrySettings';
+import { removeContainer } from '../commands/remove-container';
+import { removeImage } from '../commands/remove-image';
+import { restartContainer } from '../commands/restart-container';
+import { showLogsContainer } from '../commands/showlogs-container';
+import { startAzureCLI, startContainer, startContainerInteractive } from '../commands/start-container';
+import { stopContainer } from '../commands/stop-container';
+import { systemPrune } from '../commands/system-prune';
+import { tagImage } from '../commands/tag-image';
+import { docker } from '../commands/utils/docker-endpoint';
+import { DefaultTerminalProvider } from '../commands/utils/TerminalProvider';
+import { DockerDebugConfigProvider } from '../configureWorkspace/configDebugProvider';
+import { configure, configureApi } from '../configureWorkspace/configure';
+import { registerDebugConfigurationProvider } from '../debugging/coreclr/registerDebugger';
+import { AzureAccountWrapper } from '../explorer/deploy/azureAccountWrapper';
+import { WebAppCreator } from '../explorer/deploy/webAppCreator';
+import { DockerExplorerProvider } from '../explorer/dockerExplorerProvider';
+import { AzureImageTagNode } from '../explorer/models/azureRegistryNodes';
+import { connectCustomRegistry, disconnectCustomRegistry } from '../explorer/models/customRegistries';
+import { DockerHubImageTagNode } from '../explorer/models/dockerHubNodes';
+import { NodeBase } from '../explorer/models/nodeBase';
+import { browseAzurePortal } from '../explorer/utils/browseAzurePortal';
+import { browseDockerHub, dockerHubLogin, dockerHubLogout } from '../explorer/utils/dockerHubUtils';
 import { COMPOSE_FILE_GLOB_PATTERN, configPrefix, configurationKeys } from './constants';
-import { registerDebugConfigurationProvider } from './debugging/coreclr/registerDebugger';
 import { DockerComposeCompletionItemProvider } from './dockerCompose/dockerComposeCompletionItemProvider';
 import { DockerComposeHoverProvider } from './dockerCompose/dockerComposeHoverProvider';
 import composeVersionKeys from './dockerCompose/dockerComposeKeyInfo';
 import { DockerComposeParser } from './dockerCompose/dockerComposeParser';
-import { DockerfileCompletionItemProvider } from './dockerfile/dockerfileCompletionItemProvider';
-import DockerInspectDocumentContentProvider, { SCHEME as DOCKER_INSPECT_SCHEME } from './documentContentProviders/dockerInspect';
-import { AzureAccountWrapper } from './explorer/deploy/azureAccountWrapper';
-import { WebAppCreator } from './explorer/deploy/webAppCreator';
-import { DockerExplorerProvider } from './explorer/dockerExplorerProvider';
-import { AzureImageTagNode } from './explorer/models/azureRegistryNodes';
-import { connectCustomRegistry, disconnectCustomRegistry } from './explorer/models/customRegistries';
-import { DockerHubImageTagNode } from './explorer/models/dockerHubNodes';
-import { NodeBase } from './explorer/models/nodeBase';
-import { browseAzurePortal } from './explorer/utils/browseAzurePortal';
-import { browseDockerHub, dockerHubLogin, dockerHubLogout } from './explorer/utils/dockerHubUtils';
+import { DockerfileCompletionItemProvider } from './dockerfileCompletionItemProvider';
+import DockerInspectDocumentContentProvider, { SCHEME as DOCKER_INSPECT_SCHEME } from './dockerInspect';
 import { DefaultImageGrouping, ext, ImageGrouping } from './extensionVariables';
-import { addUserAgent } from './src/utils/addUserAgent';
-import { AzureUtilityManager } from './src/utils/azureUtilityManager';
-import { getTrustedCertificates } from './src/utils/getTrustedCertificates';
-import { Keytar } from './src/utils/keytar';
-import { wrapError } from './src/utils/wrapError';
+import { addUserAgent } from './utils/addUserAgent';
+import { AzureUtilityManager } from './utils/azureUtilityManager';
+import { getTrustedCertificates } from './utils/getTrustedCertificates';
+import { Keytar } from './utils/keytar';
+import { wrapError } from './utils/wrapError';
 
 export type KeyInfo = { [keyName: string]: string };
