+++ src/utils/osVersion.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as semver from 'semver';
-import { ext } from '../../extensionVariables';
+import { ext } from '../extensionVariables';
 
 // Minimum Windows RS3 version number
 const windows10RS3MinVersion = '10.0.16299';
